move marker to separate file; parser now can return markers directly
